Move Utf8String to core AsmResolver package.
